requirements-poc.txt
aiortc==1.9.0
av==10.0.0
aiohttp==3.8.1
numpy==1.25.0
Pillow==9.5.0
